+++ CHANGELOG.md
@@ -6,7 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-
+## [0.1.1] - 2020-01-28
+### Added
+- Expand timestamping capabilities, i.e. `timestamp-mode` property that can take the following forms: `ignore`, `main`, `all`, `k4a_common` and `k4a_individual`
+- Support for MJPG
+### Modified
+- Fix streaming from Playback.
 
 ## [0.1.0] - 2020-01-24
 ### Added